Add findood script and cpandists, which it requires.
cpandists prints a list of all CPAN distribution names and versions.
Only the latest versions are printed. Not foolproof but works so far.

findood receives the output of findrepopkgs and prints out the
names of any out of date packages, using the output of cpandists
saved to a filed called "dists" to compare versions.